requirements.txt
# ============================================
# Core Dependencies
# ============================================
numpy==2.2.6
python-dotenv==1.1.1

# ============================================
# Data Preprocessing (src/preprocessing/)
# ============================================
# 이미지 처리
pillow==11.3.0              # 이미지 로드, 증강 (data_loader.py, data_augmentation.py)
opencv-python==4.12.0.88    # 이미지 처리 보조 (data_loader.py)

# 진행률 표시
tqdm==4.67.1                # 전처리 진행률 표시 (pipeline.py)

# ============================================
# VLM & AI Model
# ============================================
openai==2.5.0               # OpenAI API (reasoning 레이어)
torch==2.8.0                # PyTorch
torchvision==0.23.0         # 비전 모델

# Object Detection
ultralytics==8.3.212        # YOLO (monitoring 레이어)
ultralytics-thop==2.0.17

# ============================================
# Data Analysis & Visualization
# ============================================
matplotlib==3.10.7
scipy==1.15.3
polars==1.34.0
polars-runtime-32==1.34.0

# ============================================
# Jupyter Notebook
# ============================================
jupyter==1.1.1
jupyterlab==4.4.9
notebook==7.4.7
ipykernel==6.30.1
ipython==8.37.0

# Jupyter 지원 라이브러리
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
notebook_shim==0.2.4

# Notebook 렌더링
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
mistune==3.1.4
pandocfilters==1.5.1

# ============================================
# Utilities
# ============================================
requests==2.32.5
pydantic==2.12.3
pydantic_core==2.41.4
PyYAML==6.0.3

# ============================================
# Supporting Libraries
# ============================================
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
fonttools==4.60.1
fqdn==1.5.1
fsspec==2025.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lark==1.3.0
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
overrides==7.7.0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
platformdirs==4.5.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pyzmq==27.1.0
referencing==0.36.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
Send2Trash==1.8.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
sympy==1.14.0
terminado==0.18.1
tinycss2==1.4.0
tomli==2.3.0
tornado==6.5.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20251008
typing-inspection==0.4.2
typing_extensions==4.15.0
uri-template==1.3.0
urllib3==2.5.0
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
